Improve debugging info for interrupts
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/trunk/libsidplayfp@2594 75748be7-0387-4326-94c7-f41ce0883d2c

+++ src/c64/CPU/mos6510.h
@@ -150,7 +150,7 @@ private:
 
 #ifdef DEBUG
     // Debug info
-    uint_least16_t instrStartPC;
+    int_least32_t instrStartPC;
     uint_least16_t instrOperand;
 
     FILE *m_fdbg;
